Consistent module naming and snake_case file-names - renamed module 'border' to 'borders' - renamed multiple files to use snake_case
